+++ src/engine/graphics/device.cpp
@@ -74,7 +74,7 @@ static bool isDeviceSuitable(const VkPhysicalDevice& device, const frametech::gr
 
 static void listAvailableExtensions(const VkPhysicalDevice& physical_device)
 {
-    uint32_t available_extensions_count;
+    uint32 available_extensions_count;
     vkEnumerateDeviceExtensionProperties(physical_device, nullptr, &available_extensions_count, nullptr);
 
     std::vector<VkExtensionProperties> available_extensions;
@@ -126,16 +126,16 @@ frametech::graphics::Device::~Device()
     Log("< Destroying the physical, and logical, devices...");
 }
 
-uint32_t frametech::graphics::Device::getNumberDevices() const
+uint32 frametech::graphics::Device::getNumberDevices() const
 {
-    uint32_t device_count{};
+    uint32 device_count{};
     vkEnumeratePhysicalDevices(frametech::Engine::getInstance()->m_graphics_instance, &device_count, nullptr);
     return device_count;
 }
 
 ftstd::VResult frametech::graphics::Device::listDevices(const frametech::graphics::DeviceSupportsOptions& options)
 {
-    uint32_t device_count{};
+    uint32 device_count{};
     vkEnumeratePhysicalDevices(frametech::Engine::getInstance()->m_graphics_instance, &device_count, nullptr);
     if (device_count == 0)
     {
@@ -163,28 +163,28 @@ bool frametech::graphics::Device::isInitialized() const
     return VK_NULL_HANDLE != m_physical_device;
 }
 
-ftstd::Result<uint32_t> frametech::graphics::Device::getQueueFamilies()
+ftstd::Result<uint32> frametech::graphics::Device::getQueueFamilies()
 {
     assert(isInitialized());
     if (!isInitialized())
     {
-        return ftstd::Result<uint32_t>::Error((char*)"The physical device has not been setup");
+        return ftstd::Result<uint32>::Error((char*)"The physical device has not been setup");
     }
-    uint32_t total_queue_families = 0;
+    uint32 total_queue_families = 0;
 
     vkGetPhysicalDeviceQueueFamilyProperties(m_physical_device, &total_queue_families, nullptr);
     if (total_queue_families == 0)
     {
         Log("No queue families for the selected physical device");
-        return ftstd::Result<uint32_t>::Ok(total_queue_families);
+        return ftstd::Result<uint32>::Ok(total_queue_families);
     }
     std::vector<VkQueueFamilyProperties> found_queue_families(total_queue_families);
-    m_queue_support.resize((size_t)total_queue_families);
-    m_queue_states.resize((size_t)total_queue_families);
+    m_queue_support.resize((int)total_queue_families);
+    m_queue_states.resize((int)total_queue_families);
     vkGetPhysicalDeviceQueueFamilyProperties(m_physical_device, &total_queue_families, found_queue_families.data());
 
     // Constructs the set of internal queues
-    for (uint32_t i = 0; i < total_queue_families; i++)
+    for (uint32 i = 0; i < total_queue_families; i++)
     {
         Log("\t> checking queue family %d", i);
 
@@ -225,10 +225,10 @@ ftstd::Result<uint32_t> frametech::graphics::Device::getQueueFamilies()
     }
     if (!supports_graphics || !supports_present || !supports_transfert)
     {
-        return ftstd::Result<uint32_t>::Error((char*)"did not found any queue that support our requirements");
+        return ftstd::Result<uint32>::Error((char*)"did not found any queue that support our requirements");
     }
 
-    return ftstd::Result<uint32_t>::Ok(total_queue_families);
+    return ftstd::Result<uint32>::Ok(total_queue_families);
 }
 
 ftstd::VResult frametech::graphics::Device::createLogicalDevice()
@@ -249,13 +249,13 @@ ftstd::VResult frametech::graphics::Device::createLogicalDevice()
     int took_indices[3] = {-1, -1, -1};
     // Influences the scheduling of command buffer execution (1.0 is the max priority value)
     // Required, even for a single queue
-    float queue_priority = 1.0f;
+    f32 queue_priority = 1.0f;
 
-    size_t queue_index = 0;
+    int queue_index = 0;
     // TODO: initialize and set the present queue here
     for (const SupportFeatures supported_flag : supported_flags)
     {
-        uint32_t first_index = 0;
+        uint32 first_index = 0;
         for (int i = 0; i < m_queue_support.size(); i++)
         {
             first_index = i;
@@ -309,9 +309,9 @@ ftstd::VResult frametech::graphics::Device::createLogicalDevice()
     // Initializes the logical device
     VkDeviceCreateInfo logical_device_create_info{
         .sType = VK_STRUCTURE_TYPE_DEVICE_CREATE_INFO,
-        .queueCreateInfoCount = static_cast<uint32_t>(queues.size()),
+        .queueCreateInfoCount = static_cast<uint32>(queues.size()),
         .pQueueCreateInfos = queues.data(),
-        .enabledExtensionCount = static_cast<uint32_t>(REQUIRED_EXTENSIONS.size()),
+        .enabledExtensionCount = static_cast<uint32>(REQUIRED_EXTENSIONS.size()),
         .ppEnabledExtensionNames = REQUIRED_EXTENSIONS.data(),
         .pEnabledFeatures = &device_features,
     };
