+++ src/engine/graphics/texture.cpp
@@ -53,9 +53,9 @@ ftstd::VResult frametech::engine::graphics::Texture::createImage(
         .imageType = is_1D_texture ? VK_IMAGE_TYPE_1D : VK_IMAGE_TYPE_2D,
         .format = texture_format,
         .extent = {
-            .width = static_cast<uint32_t>(m_width),
-            .height = static_cast<uint32_t>(m_height),
-            .depth = is_1D_texture ? static_cast<uint32_t>(0) : static_cast<uint32_t>(1), // FIXME !!! Big errors will happen for 3D textures - handle the depth via the image load function,
+            .width = static_cast<uint32>(m_width),
+            .height = static_cast<uint32>(m_height),
+            .depth = is_1D_texture ? static_cast<uint32>(0) : static_cast<uint32>(1), // FIXME !!! Big errors will happen for 3D textures - handle the depth via the image load function,
         },
         .mipLevels = 1,
         .arrayLayers = 1,
@@ -186,7 +186,7 @@ ftstd::VResult frametech::engine::graphics::Texture::setup(
     // Now, fill the staging buffer with the actual data
     void* data;
     vmaMapMemory(resource_allocator, staging_buffer_allocation, &data);
-    memcpy(data, image_data, static_cast<size_t>(texture_size));
+    memcpy(data, image_data, static_cast<int>(texture_size));
     vmaUnmapMemory(resource_allocator, staging_buffer_allocation);
 
     // No need the raw data, as we mapped the data in VMA
@@ -217,7 +217,7 @@ ftstd::VResult frametech::engine::graphics::Texture::setup(
     auto transfert_command_buffer = frametech::Engine::getInstance()->m_render->getTransfertCommand();
     VkCommandPool* transfert_command_pool = transfert_command_buffer->getPool();
 
-    const uint32_t transfert_queue_family_index = transfert_command_buffer->m_queue_family_index_created_with;
+    const uint32 transfert_queue_family_index = transfert_command_buffer->m_queue_family_index_created_with;
 
     {
         frametech::graphics::Command command_buffer(*transfert_command_pool);
@@ -287,8 +287,8 @@ frametech::engine::graphics::DepthTexture::~DepthTexture()
     }
 }
 
-ftstd::VResult frametech::engine::graphics::DepthTexture::createImage(const uint32_t image_height,
-                                                                      const uint32_t image_width,
+ftstd::VResult frametech::engine::graphics::DepthTexture::createImage(const uint32 image_height,
+                                                                      const uint32 image_width,
                                                                       const VkFormat texture_format,
                                                                       const VkImageTiling tiling,
                                                                       const VkImageUsageFlags usage_flags) noexcept
